tons of NGit methods and ExtensionMethods

+++ UnitTests/UnitTests.FluentSharp.NGit/Helper/Temp_Repo.cs
@@ -1,18 +1,26 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using FluentSharp;
+﻿using FluentSharp;
 using FluentSharp.ExtensionMethods;
 using NUnit.Framework;
-using O2.DotNetWrappers.Windows;
 
 namespace UnitTests.FluentSharp_NGit
 {
     public class Temp_Repo
     {
-        public API_NGit nGit;
-        public string repoPath;
+        public API_NGit nGit        { get; set; }
+        public string   repoPath    { get; set; }
+        public bool     deleteRepo  { get; set; }
+
+        public Temp_Repo()
+        {
+            deleteRepo = true;
+        }
+
+        public Temp_Repo(Temp_Repo tempRepo)
+        {
+            nGit        = tempRepo.nGit;
+            repoPath    = tempRepo.repoPath;
+            deleteRepo  = tempRepo.deleteRepo;
+        }
 
         [SetUp]
         public void SetUp()
@@ -22,15 +30,24 @@ namespace UnitTests.FluentSharp_NGit
             Assert.IsEmpty(repoPath.files()          , " repoPath should be empty");
             Assert.IsEmpty(repoPath.dirs()           , " repoPath should be empty");
             nGit = repoPath.git_Init();
+
+            Assert.IsTrue(repoPath.isGitRepository());
+            Assert.IsNotNull(nGit);
+            Assert.IsNotNull(nGit.Git);
+            Assert.IsNotNull(nGit.Repository);
+
         }
 
         [TearDown]
         public void TearDown()
         {
-            Assert.IsTrue(repoPath.dirExists());
-            nGit.close()
-                .delete_Repository_And_Files();            
-            Assert.IsFalse(repoPath.dirExists());
+            if (deleteRepo)
+            {
+                Assert.IsTrue(repoPath.dirExists());
+                nGit.close()
+                    .delete_Repository_And_Files();
+                Assert.IsFalse(repoPath.dirExists());
+            }
         }
 
         [Test]
